Type error in stream_id
When no data was found, the stream id was set to 'nodata'. This turns
the id in a character type, instead of integer. Furter on in the code
the data was bind using bind_rows. When the stream data both contains
character and integer this results in an error.

With this change, the stream id is set to NA in case of no data